+++ doc/storage/conf.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: Copyright 2018-2019, 2022-2025 Arm Limited and/or its affiliates
+# SPDX-FileCopyrightText: Copyright 2018-2019, 2022-2026 Arm Limited and/or its affiliates
 # SPDX-License-Identifier: CC-BY-SA-4.0 AND LicenseRef-Patent-license
 
 # PSA Certified API document configuration
@@ -8,42 +8,34 @@
 
 doc_info = {
     # Document template
-    'template': 'psa-api-2025',
+    'template': 'psa-api-2026',
 
     # Document title, MANDATORY
     'title': 'PSA Certified\nSecure Storage API',
     'author': 'Arm Limited',
 
     # Document copyright date, default to year of 'date'
-    'copyright_date': '2018-2019, 2022-2025',
+    'copyright_date': '2018-2019, 2022-2026',
     'copyright': 'Arm Limited and/or its affiliates',
 
     # Document identifier, marked as open issue if not provided
-    'doc_id': 'IHI 0087',
+    'doc_id': 'GPD_SPE_087',
 
     # The short X.Y version. MANDATORY
     'version': '1.0',
-    # Document quality status, marked as open issue if not provided
-    'quality': 'REL',
-    # Document issue number (within that version and quality status)
-    # Marked as open issue if not provided
+    # Document maintenance revision
     'issue_no': 4,
-    # Identifies the sequence number of a release candidate of the same issue
-    # default to None
-    #'release_candidate': 1,
-    # Draft status - use this to indicate the document is not ready for publication
-    #'draft': True,
-
-    # Document confidentiality. Must be either Non-confidential or Confidential
-    # Marked as open issue if not provided
-    'confidentiality': 'Non-confidential',
+    # Document draft revision
+    'draft': 1,
+    # Document status
+    'status': 'DFT',
 
     # Id of the legal notice for this document
     # Marked as open issue if not provided
-    'license': 'psa-certified-api-license',
+    #'license': 'psa-certified-api-license',
 
     # Document date, default to build date
-    'date': '23/09/2025',
+    'date': 'September 2025',
 
     # psa_spec: default header file for API definitions
     # default to None, and can be set in documentation source
@@ -54,9 +46,6 @@ doc_info = {
     #    1 : Primary API elements
     #    2 : Sub-elements of API - parameters, fields, values
     'header_doxygen': 2,
-
-    # Declare a watermark for the PDF output
-    #'watermark': 'DRAFT',
 
     # Optional ordering of return error values
     # This list is used to create a standard ordering of return value responses